bugfix commandlineparser tmax to double
Former-commit-id: 6c8dde015a11702c899966d4d458e79394f10491 [formerly a8ce18bbe7c9f8787b5f4a43f618b60d007d2e1a]
Former-commit-id: 830459460b9ff40727352960bb4a955773cc1941

+++ src/library/natrium/utilities/CommandLineParser.cpp
@@ -383,7 +383,7 @@ void CommandLineParser::applyToSolverConfiguration(SolverConfiguration& cfg) {
 
 	// simulation end time
 	if (hasArgument("tmax")) {
-		int tmax = getArgument<double>("tmax");
+		double tmax = getArgument<double>("tmax");
 		cfg.setSimulationEndTime(tmax);
 		LOG(BASIC) << "Output simulation end time set to " << tmax
 				<< " via command line" << endl;
